+++ src/main/java/lumien/randomthings/config/Numbers.java
@@ -51,4 +51,7 @@ public class Numbers
 
 	@ConfigOption(category = CATEGORY, name = "MaxBiomeCapsuleCapacity",
 			comment = "The maximum") public static int MAX_BIOME_CAPSULE_CAPACITY = 256;
+
+	@ConfigOption(category = CATEGORY, name = "PotionVaporizerMaxBlocks", comment = "The maximum number of blocks a Potion Vaporizer can affect")
+	public static int POTION_VAPORIZER_MAX_BLOCKS = 100;
 }